restructure validation dt 100 script
