+++ runs/validation_prerun.jl
@@ -6,24 +6,21 @@ module ValidationPrerun
 
 using Unitful
 using CarboKitten
-using DelimitedFiles
-using DataFrames
 using Interpolations
 using CairoMakie
 using CarboKitten.Visualization: sediment_profile, summary_plot
 using CarboKitten.Export: read_slice, read_volume, write_csv
 using Tables
 using CarboKitten.Boxes: Periodic
-const TAG = "alcap-validation"
-const FILEPATH = "data/Morley_2021.txt"
-const DATAFILE = "data/validation_prerun_topography.csv"
-const OUTPUT_FILE = "data/validationprerun.h5"
+
+const TAG = "validation_prerun"
+
+const OUTPUT_FILE = "data/$(TAG).h5"
 const PERIOD = 0.2 * u"Myr"
 const AMPLITUDE = 30.0u"m"
 const BASE = 30.0u"m"
-const GRID_SIZE = (140, 100)
-const PHYS_SCALE = 50u"m"
 
+include("runs/ValidationConstants.jl")
 
 function dome_topography(x,y)
     X_DIM = GRID_SIZE[1] * PHYS_SCALE
@@ -39,49 +36,9 @@ function dome_topography(x,y)
     end
 end
 
-
-facies(feedback) = [
-    ALCAP.Facies(
-        viability_range = (4, 10),
-        activation_range = (6, 10),
-        active = true,
-        production = BenthicProduction(
-            maximum_growth_rate=1800u"m/Myr",
-            extinction_coefficient=0.6u"m^-1",
-            saturation_intensity=60u"W/m^2"),
-        transport_coefficient=2.0u"m/yr",
-        name="euphotic"),
-    ALCAP.Facies(
-        viability_range = (4, 10),
-        activation_range = (6, 10),
-        active = true,
-        production = BenthicProduction(
-            maximum_growth_rate=800u"m/Myr",
-            extinction_coefficient=0.1u"m^-1",
-            saturation_intensity=60u"W/m^2"),
-#        minimum_production=feedback ? 0.1u"m/Myr" : nothing,
-        transport_coefficient=2.0u"m/yr",
-        name="oligophotic"),
-    ALCAP.Facies(
-        viability_range = (4, 10),
-        activation_range = (6, 10),
-        active = false,
-        production = PelagicProduction(
-            maximum_growth_rate=8u"1/Myr",
-            extinction_coefficient=0.6u"m^-1",
-            saturation_intensity=60u"W/m^2",
-            maximum_production_depth=50u"m"),
-        transport_coefficient=2.0u"m/yr",
-        name="pelagic"),
-    ALCAP.Facies(
-        active=false,
-        transport_coefficient=2.0u"m/yr",
-        name="oligophotic transported")
-]
-
 input(feedback) = ALCAP.Input(
     tag=TAG,
-    box=CarboKitten.Box{Periodic{2}}(grid_size=GRID_SIZE, phys_scale=PHYS_SCALE),
+    box=VALIDATION_BOX,
     time=TimeProperties(
         Δt=0.0001u"Myr",
         steps=5000
@@ -100,7 +57,7 @@ input(feedback) = ALCAP.Input(
     sediment_buffer_size=50,
     depositional_resolution=0.5u"m",
     lithification_time = 50.0u"yr",
-    disintegration_transfer = f -> stack((0.0.*f[1,:,:], 0.0.*f[2,:,:], 0.0.*f[3,:,:],
+    disintegration_transfer = f -> stack((f[1,:,:], 0.0.*f[2,:,:], f[3,:,:],
                                       f[1,:,:].+f[4,:,:]), dims=1),
     facies=facies(feedback))
 
@@ -108,22 +65,10 @@ function main()
     run_model(Model{ALCAP}, input(true), "$OUTPUT_FILE")
 end
 
-function save_final_topography(prerun_filename)
-    header, data = read_volume(prerun_filename, :topography)
-
-    t = header.axes.t
-    h0 = header.initial_topography
-    subsidence = header.subsidence_rate * (t[end] - t[1])
-    delta_h = data.sediment_thickness[:, :, end]
-    h = h0 .+ delta_h .- subsidence
-
-    write_csv(DATAFILE, h |> in_units_of(u"m") |> Tables.table)
-end
-
 function preplot()
     header, data = read_slice("$OUTPUT_FILE", :profile)
 	fig = sediment_profile(header, data, show_unconformities = false)
-    save("md/fig/validation_prerun.png", fig)
+    save("md/fig/$(TAG).png", fig)
 end
 end
